Split App event-wiring into a partial class file
App.axaml.cs had grown to ~4,650 lines. Move the 11 Wire*Events /
WireIdleDetection methods into a new App.EventWiring.cs as members of the
same partial App class. This is a pure relocation: no method bodies,
signatures, call sites, or startup ordering changed, and the static App.*
surface the rest of the app depends on is untouched. App.axaml.cs drops to
~3,060 lines.
